refactor: extract god-file theme UI into ChatThemePanel and ThemeBorderResolver
- Extract _buildThemeSection() from chat_settings_dialog.dart → ChatThemePanel widget
- Extract border/theme resolution from message_bubble.dart → ThemeBorderResolver
- Add resolveThemeColors() to ChatThemeSettings.tsx for WebUI parity
- Apply theme overrides as CSS vars on live chat container (ChatPage.tsx)
- Both Dart files now at/below Rawhide line counts
