+++ web_ui/src/styles.css
@@ -298,7 +298,7 @@ button.ghost { background: transparent; border-color: var(--border); }
 .char-meta { font-size: 0.7rem; color: var(--muted); }
 
 /* ── Chat ─────────────────────────────────────────────────────── */
-.chat-view { display: flex; flex-direction: column; height: 100%; position: relative; }
+.chat-view { display: flex; flex-direction: column; height: 100%; position: relative; font-family: var(--chat-font, inherit); }
 .chat-header {
   display: flex; align-items: center; justify-content: space-between;
   padding: 12px 16px; font-weight: 600; border-bottom: 1px solid var(--border);
